fix: consolidate types files, clean up imports
