+++ models/attribute-definition/src/attribute-definition-draft/generator.ts
@@ -2,7 +2,7 @@ import * as AttributeType from '@commercetools-test-data/attribute-type';
 import { LocalizedString } from '@commercetools-test-data/commons';
 import { fake, Generator } from '@commercetools-test-data/core';
 import { attributeConstraints, inputHints } from '../constants';
-import { TAttributeDefinitionDraft } from './types';
+import { TAttributeDefinitionDraft } from '../types';
 
 // https://docs.commercetools.com/api/projects/productTypes#attributedefinitiondraft
